requirements/common.txt
cryptography==37.0.2
cx-Oracle==8.3.0
Flask==2.0.0
Jinja2==3.0.0
MarkupSafe==2.1.1
itsdangerous==2.0.0
Werkzeug==2.0.0
Flask-APScheduler==1.11.0
Flask-Babel==1.0.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
jaconv==0.2.4
numpy==1.19.3
psycopg2==2.9.5
pymssql==2.2.11
PyMySQL==1.0.2
ruamel.yaml==0.16.5
markdown2==2.3.9
pytz==2021.3
APScheduler==3.7.0
SQLAlchemy==1.3.13
pandas==1.2.5
scipy==1.7.1
alembic==1.4.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
marshmallow==3.9.1
WTForms==2.3.3
Flask-WTF==0.14.3
WTForms-JSON==0.3.5
simplejson==3.17.2
joblib==1.2.0
scikit-learn==0.24.2
chardet==3.0.4
# detect-delimiter==0.1.1
Flask-Compress==1.10.1
group-lasso==1.5.0
# mecab-python3==1.0.5
unidic-lite==1.0.8
fugashi==1.1.2
cutlet==0.1.19
grpcio==1.51.1
protobuf==3.20
redis==4.4.0
# redis-lru==0.1.2
# cachetools==4.2.4
anyascii==0.3.0
# dask==2023.2.0
langdetect==1.0.9
pyarrow==9.0.0
lz4==4.0.2
waitress==2.1.2
# requests==2.28.2
python-dateutil~=2.8.2
Flask-Assets==2.0
orjson==3.9.1
# pywin32==306
typing-extensions==4.7.1
pydantic==2.2
# more-itertools==10.1.0
